python/requirements.txt
bleach==6.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
imbalanced-learn==0.13.0
joblib==1.5.1
kaggle==1.7.4.5
numpy==1.26.4
pandas==2.3.0
protobuf==6.31.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.3
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
